网络防火墙codex清理

+++ services/netfirewallmanager/src/netfirewall_db_helper.cpp
@@ -91,7 +91,7 @@ bool NetFirewallDbHelper::BlobToDomainList(const std::vector<uint8_t> &blob, std
         }
         const uint8_t *sizePtr = &blob[i];
         uint16_t size = *((uint16_t *)sizePtr);
-        int index = i + lenSize;
+        size_t index = i + lenSize;
         if (index >= blobSize || (blobSize - index) < size) {
             return true;
         }
@@ -823,7 +823,7 @@ int32_t NetFirewallDbHelper::AddInterceptRecord(const int32_t userId, std::vecto
     rdbPredicates.BeginWrap()->EqualTo(NET_FIREWALL_USER_ID, std::to_string(userId))->EndWrap();
     firewallDatabase_->Count(currentRows, rdbPredicates);
     // Aging by number, record up to 1000 pieces of data
-    size_t size = records.size();
+    int64_t size = static_cast<int64_t>(records.size());
     int64_t leftRows = RECORD_MAX_DATA_NUM - currentRows;
     if (leftRows < size) {
         std::string whereClause("id in (select id from ");
